main_renderer.dart: stop drawing dashed lines at the right of main k_chart, because we are not drawing solid block there anymore
The solid block was drawn at the right side of main k_chart to cover the graph, that shows under numbers if you scroll fast. Right now we are ok with that amd by removing this opaque block we fixed another issue (can't remember, but related to graph drawing)

Dashed lines was needed because opaque cover hide it. Now those line duplicated and looks ugly.

https://dev.azure.com/DueDEX/Mex/_workitems/edit/1835